# Improved the parsing performance of bitmap fonts