Test Environment:
PC:Lenovo T420i
System:Window 8.1

My test case:
```cpp
static std::string fontFile = "fonts/markerFelt.fnt";
auto startTime = utils::gettime();
for (int i = 0; i < 50; ++i)
{
        auto bmfont = new (std::nothrow) BMFontConfiguration();
        bmfont->initWithFNTfile(fontFile);
        bmfont->release();
}
auto dt = utils::gettime() - startTime;
```

The most recent test result with markerFelt.fnt[146kb]：
1.the latest version of cocos2d-x v3 branch：35.32s
3.applied my patch:8.73s

The most recent test result with arial16.fnt[22kb]：
1.the latest version of cocos2d-x v3 branch：1.292s
3.applied my patch:0.267s