@minggo 2015-01-26
Thanks @Dhilan007 
Should test on iOS and Android.