@Dhilan007 2015-01-26
Update：
Test case:
```cpp
static std::string fontFile = "fonts/markerFelt.fnt";
auto startTime = utils::gettime();
for (int i = 0; i < 50; ++i)
{
        auto bmfont = new (std::nothrow) BMFontConfiguration();
        bmfont->initWithFNTfile(fontFile);
        bmfont->release();
}
auto dt = utils::gettime() - startTime;
```
Test on android[Nexus 9]:
1.the latest version of cocos2d-x v3 branch：7.53s
2.applied my patch:0.4990s

Test on ios[iphone 5s]:
1.the latest version of cocos2d-x v3 branch：14.39s
2.applied my patch:0.3019s